Refactor localstorage, auth, add mock store
