+++ src/utils/auth.js
@@ -1,0 +1,12 @@
+import { storage } from './localStorage';
+
+const AUTH_KEY = 'auth';
+
+export const auth = {
+  set(authStatus) {
+    storage.set(AUTH_KEY, authStatus);
+  },
+  isAuth() {
+    return storage.get(AUTH_KEY)
+  },
+};
