+++ src/utils/localStorage.js
@@ -1,0 +1,8 @@
+export const storage = {
+  set(key, value) {
+    localStorage.setItem(key, JSON.stringify(value));
+  },
+  get(key) {
+    return JSON.parse(localStorage.getItem(key));
+  },
+};
